+++ lib/runger_style/cops/default/monkeypatches/multiline_element_line_breaks.rb
@@ -4,16 +4,16 @@ module RungerStyle::MultilineElementLineBreaksPatches
   private
 
   def check_line_breaks(node, children, ignore_last: false)
-    return if single_line?(node, children, ignore_last:)
-
-    last_seen_line = -1
-    children.each do |child|
-      if last_seen_line >= child.first_line
-        add_offense(child) do |corrector|
-          RuboCop::Cop::EmptyLineCorrector.insert_before(corrector, child)
+    unless single_line?(node, children, ignore_last:)
+      last_seen_line = -1
+      children.each do |child|
+        if last_seen_line >= child.first_line
+          add_offense(child) do |corrector|
+            RuboCop::Cop::EmptyLineCorrector.insert_before(corrector, child)
+          end
+        else
+          last_seen_line = child.last_line
         end
-      else
-        last_seen_line = child.last_line
       end
     end
   end